more refactoring and changes to dependency
