+++ optml/__init__.py
@@ -1,6 +1,5 @@
 from optml import models
-from optml.bayesian_optimization import bayesian_optimizer
-from optml import bayesian_optimization
+from optml import bayesian_optimizer
 from optml import genetic_optimizer
 from optml import gridsearch_optimizer
 from optml import random_search
