+++ optml/bayesian_optimizer/__init__.py
@@ -1,0 +1,1 @@
+from bayesian_optimizer import BayesianOptimizer
