+++ optml/bayesian_optimizer/gp_categorical.py
@@ -1,3 +1,4 @@
+import numpy as np
 from operator import itemgetter
 from scipy.linalg import cholesky, cho_solve, solve_triangular
